fix: fix jpa projection

+++ main/src/main/java/ru/practicum/main/service/event/EventServiceImpl.java
@@ -255,13 +255,14 @@ public class EventServiceImpl implements EventService {
 
     private List<Event> getConfirmedRequestsCountForEvents(List<Event> eventList) {
         if (!eventList.isEmpty()) {
-            Map<Long, Integer> confirmRequestsCount = eventRepository.getConfirmedRequestCountForEvents(
+            Map<Long, Long> confirmRequestsCount = eventRepository.getConfirmedRequestCountForEvents(
                             eventList.stream().map(Event::getId).collect(Collectors.toList())).stream()
                     .collect(Collectors.toMap(EventsRequestCount::getEventId, EventsRequestCount::getReqCount));
 
-            return eventList.stream().map(e -> e.withConfirmedRequests(confirmRequestsCount.get(e.getId())))
+            return eventList.stream().map(e -> e.withConfirmedRequests(confirmRequestsCount.get(e.getId()).intValue()))
                     .collect(Collectors.toList());
         }
+
         return Collections.emptyList();
     }
 }
